finished reset-team-info and delete-team/ added MatchSetup/ fixed some bugs

+++ src/Commands/BotAdmin/add-game.js
@@ -2,7 +2,7 @@ const { SlashCommandBuilder, PermissionFlagsBits, Client} = require('discord.js'
 const fs = require('fs');
 const games = fs.readFileSync('games.json', 'utf8');
 const gameChoices = JSON.parse(games);
-const { reloadTeamsAndGamesCommands } = require("../../Handlers/reload-commands");
+const { reloadTeamsAndGamesCommands } = require("../../Handlers/reload-global-commands");
 
 
 module.exports = {
@@ -19,18 +19,25 @@ module.exports = {
     async execute(interaction, client) {
         const gameName = interaction.options.getString('game_name');
         let gameObject = {"name": gameName, "value": gameName};
+        let gameExists = gameChoices.some(game => game.name === gameName);
+        if (gameExists){
+            await interaction.reply({
+                content: `the game: ${gameName}, is already in the json file no need to add it again`,
+                ephemeral: true
+            });
+            return;
+        }
         try{
             gameChoices.push(gameObject)
             fs.writeFileSync('games.json', JSON.stringify(gameChoices, null, 2));
             reloadTeamsAndGamesCommands(client)
+            await interaction.reply({
+                content: `game ${gameName}, successfully added to the games file, and commands were reloaded.`, 
+                ephemeral: true
+            });
         } catch(error) {
             console.log(`there was an error: ${error}`)
             return;
         }
-        await interaction.reply({
-            content: `game ${gameName}, successfully added to the games file, and commands were reloaded.`, 
-            ephemeral: true
-        })
-    
     }
 }
